[LLM-ARCH] Implement RAMpart core library
Complete implementation of all library modules:

- rp_thread.c: Platform-independent mutex and thread ID handling
- rp_wipe.c: Multi-pass secure memory wiping (0x00, 0xFF, 0xAA)
- rp_crypto.c: 16-round Feistel cipher with S-box substitution
- rp_block.c: Block header, guard band initialization and validation
- rp_pool.c: Worst-fit allocator with block splitting and coalescing
- rampart.c: Public API implementation

All 12 core tests pass with 72 assertions.

+++ h/internal/rp_types.h
@@ -286,7 +286,7 @@ typedef struct rp_pool_header_s {
  * @param x     Value to align
  * @param a     Alignment (must be power of 2)
  */
-#define RP_ALIGN_UP(x, a) (((x) + ((a) - 1)) & ~((a) - 1))
+#define RP_ALIGN_UP(x, a) (((x) + ((size_t)(a) - 1)) & ~((size_t)(a) - 1))
 
 /**
  * @def RP_BLOCK_TO_USER
